Checkpoint support for standard renames
Standard renames are included in json file in package.
These are computed from renames available from github
that were generated from conda-forge.

There is also a function that will update the renames
files from github if the etag has changed. This should
be used before building a distribution of this package.

We probably should also allow this file to be updated
after installation, either in place in the installation
directory or in some other location
(e.g. `~/.config/whl2conda/stdrenames.json`).